ai-ml-projects/image-processing/background-remover/requirements.txt
rembg==2.0.56
pillow==10.2.0
numpy==1.24.3
opencv-python==4.9.0.80
streamlit==1.31.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
onnxruntime==1.17.0
